Fixed package-info.java for Intelij, moved DataHelper, EnvironmentHelper and Helper.
