+++ src/main/java/environment/package-info.java
@@ -1,5 +1,4 @@
 /**
  * @deprecated As of release 2.0, replaced by{@link net.itarray.automotion.tools.environment}
  */
-@Deprecated()
 package environment;
