+++ src/main/java/http/connections/package-info.java
@@ -1,5 +1,4 @@
 /**
  * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.tools.http.connections}
  */
-@Deprecated()
 package http.connections;
