+++ src/main/java/http/helpers/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * @deprecated As of release 2.0, replaced by {@link net.itarray.automotion.tools.http.helpers}
+ */
+package http.helpers;
